feat: tempmute command

+++ modules/tasks.js
@@ -8,6 +8,9 @@ let db = new Database('data.db');
 
 let taskIDs = new Array()
 
+// Get client
+//const { client } = require('@root/index.js')
+
 // Chalk
 const chalk = require('chalk');
 const infoPrefixColor = chalk.black.bgWhite
@@ -16,26 +19,30 @@ const errorPrefixColor = chalk.white.bgRed
 const urlColor = chalk.blue.underline
 const highlightColor = chalk.yellow
 
-module.exports = (type, task, userID, runAt) => {
-    db.exec('CREATE TABLE IF NOT EXISTS tasks (id SMALLINT(255), task VARCHAR(255), user_id BIGINT(255), run_at TIMESTAMP)');
-    const tasksTableSelection = db.prepare('SELECT * FROM tasks');
+module.exports = async (type, task, userID, runAt) => {
+    db.exec('CREATE TABLE IF NOT EXISTS tasks (id SMALLINT(255), task VARCHAR(255), user_id text, run_at TIMESTAMP)');
+    db.defaultSafeIntegers(true);
+    const tasksTableSelection = db.prepare('SELECT * FROM tasks')
     const tasksTableGetSelection = tasksTableSelection.all();
 
     const insert = db.prepare('INSERT INTO tasks (id, task, user_id, run_at) VALUES (?, ?, ?, ?)');
+
+    const { client } = require('@root/index.js')
 
     switch (type) {
         case 'list':
             if (tasksTableGetSelection.length > 0) {
                 console.log(infoPrefixColor(config.ConsoleStyle.Prefix.Info), `====== Aesthetic BOT Task Queue List ======`)
+                let UNIX = new Date()
+                UNIX = Math.floor(Date.now() / 1000)
+                console.log(infoPrefixColor(config.ConsoleStyle.Prefix.Info), `Current UNIX time (s): ${UNIX}`)
                 for (listTask of tasksTableGetSelection) {
-                    let date = new Date(0)
-                    date.setUTCSeconds(listTask.run_at)
                     switch (listTask.task) {
                         case 'unban':
-                            console.log(`[${listTask.id}]> Unban ${listTask.user_id} by ${date} (UNIX ${listTask.run_at})`)
+                            console.log(`[${listTask.id}]> Unban ${listTask.user_id} at UNIX ${listTask.run_at}`)
                             break
                         case 'unmute':
-                            console.log(`[${listTask.id}]> Unmute ${listTask.user_id} by ${date} (UNIX ${listTask.run_at})`)
+                            console.log(`[${listTask.id}]> Unmute ${listTask.user_id} at UNIX ${listTask.run_at}`)
                             break
                         default:
                             console.log(warnPrefixColor(config.ConsoleStyle.Prefix.Warn), `Task ${listTask.task} is not a recognised task`)
@@ -47,7 +54,7 @@ module.exports = (type, task, userID, runAt) => {
             return
         case 'create':
             if (!task || !userID || !runAt) return console.log(warnPrefixColor(config.ConsoleStyle.Prefix.Warn), `Missing arguments to create a task.`)
-            
+
             for (taskID of tasksTableGetSelection) {
                 if (!taskIDs.includes(taskID.id)) taskIDs.push(taskID.id)
             }
@@ -55,12 +62,13 @@ module.exports = (type, task, userID, runAt) => {
             while (taskIDs.includes(nextTaskID)) {
                 nextTaskID = Math.floor(Math.random() * 32767)
             }
-
+            
             insert.run(nextTaskID, `${task}`, userID, runAt)
             taskIDs.push(nextTaskID)
             return
         case 'run':
             let runDate = Math.floor(Date.now() / 1000)
+            const guild = await client.guilds.cache.get(`${config.ServerID}`)
             for (runTask of tasksTableGetSelection) {
                 if (runDate >= runTask.run_at) {
                     switch (runTask.task) {
@@ -68,15 +76,26 @@ module.exports = (type, task, userID, runAt) => {
                             db.exec(`DELETE FROM tasks WHERE ${runTask.id}`);
                             removeID = taskIDs.indexOf(runTask.id)
                             taskIDs.splice(removeID, 1, '');
-                            console.log(`[${listTask.id}]> Unban of ${listTask.user_id} executed`)
+                            console.log(`[${runTask.id}]> Unban of ${runTask.user_id} executed`)
                             break
                         case 'unmute':
+                            let muteRole = guild.roles.cache.find(role => role.name === config.Moderation.Mute.Role);
+                            if (!muteRole) return console.log('No muted role was found in the server, unmute task cannceled.')
+
+                            let targetUser = guild.members.cache.get(`${runTask.user_id}`)
+                            if (!targetUser.roles.cache.has(muteRole.id)) return
+                            
+                            targetUser.roles.remove(muteRole.id)
+
                             db.exec(`DELETE FROM tasks WHERE ${runTask.id}`);
                             removeID = taskIDs.indexOf(runTask.id)
                             taskIDs.splice(removeID, 1, '');
-                            console.log(`[${listTask.id}]> Unmute of ${listTask.user_id} executed`)
-                            break
+                            console.log(`[${runTask.id}]> Unmute of ${runTask.user_id} executed`)
+                            return
                         default:
+                            db.exec(`DELETE FROM tasks WHERE ${runTask.id}`);
+                            removeID = taskIDs.indexOf(runTask.id)
+                            taskIDs.splice(removeID, 1, '');
                             break
                     }
                 }
@@ -84,7 +103,7 @@ module.exports = (type, task, userID, runAt) => {
             return
         case 'clear':
             db.exec('DROP TABLE IF EXISTS tasks');
-            db.exec('CREATE TABLE IF NOT EXISTS tasks (id SMALLINT(255), task VARCHAR(255), user_id BIGINT(255), run_at TIMESTAMP)');
+            db.exec('CREATE TABLE IF NOT EXISTS tasks (id SMALLINT(255), task VARCHAR(255), user_id text, run_at TIMESTAMP)');
             return
         default:
             return console.log(warnPrefixColor(config.ConsoleStyle.Prefix.Warn), `${type} is not a valid task type.`)
